requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
# mysqlclient==2.2.0  # 使用pymysql替代
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7
ratelimit==2.2.1
loguru==0.7.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2